@@ -6,13 +6,12 @@ import ee.joonasvali.mirrors.scene.genetic.Genome;
 import ee.joonasvali.mirrors.scene.genetic.impl.AcceleratorGene;
 import ee.joonasvali.mirrors.scene.genetic.impl.BenderGene;
 import ee.joonasvali.mirrors.scene.genetic.util.SerializationUtil;
-import org.apache.commons.lang3.builder.EqualsBuilder;
-import org.junit.Assert;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
 
 import java.io.IOException;
+import java.lang.reflect.Field;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -22,28 +21,31 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Objects;
 
 public class SerializationUtilTest {
 
-  @Rule
-  public TemporaryFolder folder = new TemporaryFolder();
+  @TempDir
+  public Path folder;
 
   @Test
   public void singleGeneSerializedAndDeserialized() throws IOException {
-    Path temp = folder.newFolder("mirrors-test").toPath();
+    Path temp = folder.resolve("mirrors-test");
     SerializationUtil util = new SerializationUtil(temp.resolve("test-evolution"));
     Genome genome = new Genome(Collections.singletonList(new BenderGene(50, 60, 70, 80)));
     util.serialize(genome, "abc");
     Path file = temp.resolve("test-evolution").resolve("abc.json");
     Genome result = SerializationUtil.deserialize(file);
     BenderGene bender = (BenderGene) result.get(0);
-    boolean equals = EqualsBuilder.reflectionEquals(bender, new BenderGene(50, 60, 70, 80));
-    Assert.assertTrue("Bender gene not equal to expected bender gene", equals);
+    Assertions.assertEquals(50, bender.x);
+    Assertions.assertEquals(60, bender.y);
+    Assertions.assertEquals(70, bender.radius);
+    Assertions.assertEquals(80, bender.strength);
   }
 
   @Test
   public void multipleGenesSerializedAndDeserialized() throws IOException {
-    Path temp = folder.newFolder("mirrors-test").toPath();
+    Path temp = folder.resolve("mirrors-test");
     SerializationUtil util = new SerializationUtil(temp.resolve("test-evolution"));
     List<Gene> geneList = new ArrayList<>();
     geneList.add(new BenderGene(50, 60, 70, 80));
@@ -54,14 +56,14 @@ public class SerializationUtilTest {
 
     Path file = temp.resolve("test-evolution").resolve("abc.json");
     Genome result = SerializationUtil.deserialize(file);
-    Assert.assertEquals(3, result.size());
+    Assertions.assertEquals(3, result.size());
 
     assertAllElementsPresentByFieldComparison(geneList, result);
   }
 
   @Test
   public void serializePopulation() throws IOException {
-    Path temp = folder.newFolder("mirrors-test").toPath();
+    Path temp = folder.resolve("mirrors-test");
     Path evolutionDir = temp.resolve("test-evolution");
     Files.createDirectories(evolutionDir);
 
@@ -99,12 +101,12 @@ public class SerializationUtilTest {
         "]";
 
     String contents = new String(Files.readAllBytes(evolutionFile), StandardCharsets.UTF_8);
-    Assert.assertEquals(expected, contents);
+    Assertions.assertEquals(expected, contents);
   }
 
   @Test
   public void deserializePopulation() throws IOException {
-    Path temp = folder.newFolder("mirrors-test").toPath();
+    Path temp = folder.resolve("mirrors-test");
     Path evolutionDir = temp.resolve("test-evolution");
     Files.createDirectories(evolutionDir);
 
@@ -145,7 +147,7 @@ public class SerializationUtilTest {
 
   @Test
   public void overwritesExistingPopulationFile() throws IOException {
-    Path temp = folder.newFolder("mirrors-test").toPath();
+    Path temp = folder.resolve("mirrors-test");
     Path evolutionDir = temp.resolve("test-evolution");
     Files.createDirectories(evolutionDir);
 
@@ -191,7 +193,7 @@ public class SerializationUtilTest {
       Iterator<T> iterator = expectedCopy.iterator();
       while (iterator.hasNext()) {
         T candidate = iterator.next();
-        if (EqualsBuilder.reflectionEquals(candidate, obj)) {
+        if (reflectionEquals(candidate, obj)) {
           iterator.remove();
           break;
         }
@@ -201,5 +203,33 @@ public class SerializationUtilTest {
     if (!expectedCopy.isEmpty()) {
       throw new Error("Unable to find objects from provided data: " + new Gson().toJson(expectedCopy));
     }
+  }
+
+  public static boolean reflectionEquals(Object a, Object b) {
+    if (a == b) {
+      return true;
+    }
+    if (a == null || b == null) {
+      return false;
+    }
+    if (a.getClass() != b.getClass()) {
+      return false;
+    }
+
+    Field[] fields = a.getClass().getDeclaredFields();
+    for (Field field : fields) {
+      field.setAccessible(true);
+      try {
+        Object valueA = field.get(a);
+        Object valueB = field.get(b);
+        if (!Objects.equals(valueA, valueB)) {
+          return false;
+        }
+      } catch (IllegalAccessException e) {
+        throw new RuntimeException(e);
+      }
+    }
+
+    return true;
   }
 }
